moved AbstractAlgorithm outside of companion object
